[feat] extract Results component from MainCard (ONE-94)
